+++ src/systems/ec/implementations/react/storybook/.storybook/addons.js
@@ -3,4 +3,4 @@ import '@storybook/addon-knobs/register';
 import '@storybook/addon-viewport/register';
 import '@storybook/addon-a11y/register';
 import '@storybook/addon-cssresources/register';
-import './addons/code/register';
+import '../../../../../../tools/storybook-addon-code/register';
