+++ src/systems/ec/implementations/react/storybook/.storybook/config.js
@@ -4,7 +4,7 @@ import { create } from '@storybook/theming';
 import { withA11y } from '@storybook/addon-a11y';
 import { INITIAL_VIEWPORTS } from '@storybook/addon-viewport';
 import { withCssResources } from '@storybook/addon-cssresources';
-import { withCode } from './addons/code/index';
+import { withCode } from '../../../../../../tools/storybook-addon-code/index';
 
 import './ECL';
