+++ src/systems/ec/implementations/react/storybook/package.json
@@ -22,8 +22,6 @@
     "@storybook/addon-options": "5.0.11",
     "@storybook/addon-viewport": "5.0.11",
     "@storybook/addons": "5.0.11",
-    "@storybook/client-logger": "5.0.11",
-    "@storybook/components": "5.0.11",
     "@storybook/react": "5.0.11",
     "@storybook/theming": "5.0.11",
     "babel-core": "^7.0.0-bridge.0",
