+++ src/tools/storybook-addon-code/package.json
@@ -1,0 +1,15 @@
+{
+  "private": true,
+  "name": "@ecl/storybook-addon-code",
+  "version": "2.3.0",
+  "main": "index.js",
+  "dependencies": {
+    "@storybook/addons": "5.0.11",
+    "@storybook/components": "5.0.11",
+    "@storybook/theming": "5.0.11",
+    "js-beautify": "1.10.0",
+    "prop-types": "15.7.2",
+    "react": "16.8.6",
+    "react-dom": "16.8.6"
+  }
+}
